+++ packages/react-form/src/hooks/list/baselist.ts
@@ -46,6 +46,9 @@ interface BaseList<Item extends object> {
   dispatch: React.Dispatch<ListAction<Item>>;
   reset(): void;
   dirty: boolean;
+  defaultValue: Item[];
+  value: Item[];
+  newDefaultValue(newDefaultItems: Item[]): void;
 }
 
 export function useBaseList<Item extends object>(
@@ -63,7 +66,8 @@ export function useBaseList<Item extends object>(
     if (!isEqual(list, state.initial)) {
       dispatch(reinitializeAction(list));
     }
-  }, [list, state.initial, dispatch]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [list, dispatch]);
 
   const validationConfigs = useMemo(
     () =>
@@ -77,6 +81,10 @@ export function useBaseList<Item extends object>(
 
   function reset() {
     dispatch(resetListAction());
+  }
+
+  function newDefaultValue(newDefaultItems: Item[]) {
+    dispatch(reinitializeAction(newDefaultItems));
   }
 
   const handlers = useHandlers(state, dispatch, validationConfigs);
@@ -105,5 +113,13 @@ export function useBaseList<Item extends object>(
 
   const fieldsDirty = useDirty({fields});
 
-  return {fields, dispatch, reset, dirty: fieldsDirty || isBaseListDirty};
+  return {
+    fields,
+    dispatch,
+    reset,
+    dirty: fieldsDirty || isBaseListDirty,
+    defaultValue: state.initial,
+    value: listWithoutFieldStates,
+    newDefaultValue,
+  };
 }