+++ packages/react-form/src/hooks/list/dynamiclist.ts
@@ -14,6 +14,9 @@ export interface DynamicList<Item extends object> {
   moveItem(fromIndex: number, toIndex: number): void;
   reset(): void;
   dirty: boolean;
+  value: Item[];
+  defaultValue: Item[];
+  newDefaultValue(newDefaultItems: Item[]): void;
 }
 
 type FactoryFunction<Item extends object> = (
@@ -34,10 +37,15 @@ export function useDynamicList<Item extends object>(
   fieldFactory: FactoryFunction<Item>,
   validationDependencies: unknown[] = [],
 ): DynamicList<Item> {
-  const {fields, dispatch, reset, dirty} = useBaseList(
-    listOrConfig,
-    validationDependencies,
-  );
+  const {
+    fields,
+    dispatch,
+    reset,
+    dirty,
+    newDefaultValue,
+    value,
+    defaultValue,
+  } = useBaseList(listOrConfig, validationDependencies);
 
   function addItem(factoryArgument?: any) {
     const itemToAdd = fieldFactory(factoryArgument);
@@ -57,5 +65,15 @@ export function useDynamicList<Item extends object>(
     dispatch(removeFieldItemAction(index));
   }
 
-  return {fields, addItem, removeItem, moveItem, reset, dirty};
+  return {
+    fields,
+    addItem,
+    removeItem,
+    moveItem,
+    reset,
+    dirty,
+    value,
+    newDefaultValue,
+    defaultValue,
+  };
 }